+++ CHANGES.rst
@@ -12,6 +12,8 @@ Unreleased
 
 - Relaxed pins from dependencies, keeping only the lower pins. This allows docstrfmt to
   be installed alongside other packages more easily.
+- Comment blocks that do not contain new-line characters are now kept on a single output
+  line.
 
 **Removed**