+++ CHANGES.rst
@@ -7,6 +7,7 @@ Unreleased
 **Added**
 
 - Added support for Python 3.13.
+- Added support for citation role.
 
 **Changed**